requirements.txt
einops
flax
jax
jaxlib
matplotlib
ml-dtypes
ml_collections
numpy
pandas
pydantic
streamlit
streamlit-code-editor
streamlit-pydantic